+++ CHANGELOG.md
@@ -8,6 +8,7 @@ v5.8.0
 - Add FieldConvert module to perform local stability analysis for compressible flows (!1319)
 - Remove get() accessor from Array data structure (!1937)
 - Fix issue with `StdTetExp::v_LocCollapsedToLocCoord` (!1946)
+- Fix issue with `NodalTriExp::v_GetTracePhysVals` (!1951)
 
 **CI**
 -- fix CubeAllElements performance test tolerance (!1943)